Make core-drawer-panel go fast using will-change
This patch teaches core-drawer-panel about the will-change CSS property. On an
element like the #drawer, we can hint to the browser that we're going to be
changing the transform property often by setting will-change: transform.

In this particular case, that causes the drawer to always be a stacking context
so that we don't trigger a layout at the end of the drawer opening animation.
Also, marking the drawer as will-change: transform keeps the open drawer in a
composited layer, which reduces repaints and hence latency when the user begins
dragging the drawer.

Finally, I've added feature detection for whether the browser supports the
unprefixed transform property. This change improves performance in browsers
that support both the prefixed and unprefixed transform properties because we
don't redundantly set both properties.